Separating utils files for each functionality
