+++ src/utils/isTypeofSvgChar.ts
@@ -1,0 +1,5 @@
+import { SvgChar } from '@/types/font';
+
+export function isTypeofSvgChar(char: unknown): char is SvgChar {
+	return (char as SvgChar)?.svgViewBox !== undefined;
+}
